- Fixed a bug that occurred when leaving directory blank
- Bug Description -
SetDirectory() would only return when a valid path was found. Since this was done recursively it also returned any previous invalid attempts AFTER the valid one overriding the directory variable.

Wasn't sure how to properly fix the recursive loop so I made a global variable for the directory and replaced the method with a void return type.